Lowest Common Ancestor Binary Tree Problem
Currently untested, just need to write a BT and pass the target nodes as args
Attempted solution with recursion and confidant it works. Could try and move 2nd recursive function out and have a while(?) loop instead which better aligns with my understanding of JS best practices but the double recursion reads very nicely to me.

Could significantly improve current solution by adding cache of checked subtrees and not rechecking them over and over as we traverse up the nodes to the root.