Added caching to improve solution
the findLCA function reassigns the last root every unsuccessful loop and the searchChildren function has a check to avoid searching through the subtree of the last root, Means each node will only be checked once instead of n times. So the big O goes from n^2 to n.

it's initialized as null